# backend/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: YSpotify
  version: '1.0'
paths:
...
  /api/spotify/login:
    get:
      tags:
        - Spotify
      summary: Link your Client Account to Spotify.
      responses:
        '200':
          description: Returns a URL that is Used to authorize the Client to access your Spotify Account
        '400':
          description: If the Request failed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/spotify/login/callback:
    get:
      tags:
        - Spotify
      summary: Callback for Logging linking your Spotify Account to the Client.
      parameters:
        - name: code
          in: query
          description: The code used to request Tokens to Spotify
          schema:
            type: string
        - name: state
          in: query
          description: The state of the Request
          schema:
            type: string
      responses:
        '200':
          description: The Account was Linked to Spotify
        '400':
          description: If the Request failed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: If there was no User that matched the state
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/spotify/topTracksPlaylist:
    post:
      tags:
        - Spotify
      summary: Create a Spotify Playlist on your Spotify Account containing the Top Tracks of the User with the given Client ID
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  format: int32
            encoding:
              userId:
                style: form
      responses:
        '202':
          description: The Playlist was successfully Requested for Creation
        '204':
          description: If the requested User has no Top Tracks to create a Playlist with
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: If either of the Users are not Linked to Spotify
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: If a Resource was not found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: http
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]